pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drift-toolkit"
version = "0.1.0"
description = "Directed Residual Intervention for Functional Testing — persona drift security testing for LLMs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "InsideTrust"},
]
keywords = ["llm-security", "red-team", "activation-steering", "persona-drift", "ai-safety"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "transformers>=4.40",
    "accelerate>=0.30",
    "huggingface-hub>=0.23",
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
quant = [
    "bitsandbytes>=0.43",
]
web = [
    "gradio>=4.0",
    "plotly>=5.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
drift = "drift.cli:app"

[project.urls]
Homepage = "https://github.com/insidetrust/drift"
Repository = "https://github.com/insidetrust/drift"
Issues = "https://github.com/insidetrust/drift/issues"

[tool.setuptools.packages.find]
include = ["drift*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
